renamed styles.jsx to style.js
