--- debug.log
@@ -1,1 +1,0 @@
-[1228/005900.513:ERROR:directory_reader_win.cc(43)] FindFirstFile: The system cannot find the path specified. (0x3)
